rename spikeglx_to_nwb.py to ibl_recording_to_nwb.py
